+++ scripts/pack-beta.mjs
@@ -36,7 +36,7 @@ try {
     exports: {
       ...metadata.exports,
       '.': metadata.exports['./beta'],
-      './stable': metadata.exports['.'],
+      './stable': metadata.exports['./playwright'],
     },
     publishConfig: { ...metadata.publishConfig, tag: 'beta' },
   }